chore(web): cleanup components and validators
