--- packages/web/src/common/validators/grid.event.validator.ts
@@ -1,11 +1,0 @@
-import {
-  type GridEvent,
-  GridEventSchema,
-  type WebEvent,
-} from "@web/common/types/web.event.types";
-
-export const validateGridEvent = (event: WebEvent): GridEvent => {
-  const result = GridEventSchema.parse(event);
-
-  return result;
-};
