--- packages/web/src/components/DND/DropZone.tsx
@@ -1,33 +1,0 @@
-import classNames from "classnames";
-import { type HTMLAttributes, type Ref } from "react";
-
-interface Props extends HTMLAttributes<HTMLDivElement> {
-  innerRef?: Ref<HTMLDivElement>;
-  isActive: boolean;
-  isInvalid?: boolean;
-}
-
-export const DropZone = ({
-  className,
-  innerRef,
-  isActive,
-  isInvalid = false,
-  ...props
-}: Props) => {
-  return (
-    <div
-      {...props}
-      aria-invalid={isInvalid || undefined}
-      className={classNames(
-        "relative rounded-default border-2 border-dashed transition-[background-color,border-color] duration-200",
-        isInvalid
-          ? "border-error border-solid bg-error/10"
-          : isActive
-            ? "border-border bg-surface"
-            : "border-transparent bg-transparent",
-        className,
-      )}
-      ref={innerRef}
-    />
-  );
-};
