--- packages/web/src/components/Divider/types.ts
@@ -1,6 +1,0 @@
-export interface Props {
-  color?: string;
-  toggled?: boolean;
-  width?: string;
-  withAnimation?: boolean;
-}
