--- packages/web/src/components/Icons/CircleTwo.tsx
@@ -1,9 +1,0 @@
-import { type IconProps, NumberCircleTwo } from "@phosphor-icons/react";
-import { getInteractiveIconClassName } from "./icon.utils";
-
-export const CircleTwoIcon = ({ className, ...props }: IconProps) => (
-  <NumberCircleTwo
-    className={getInteractiveIconClassName(className)}
-    {...props}
-  />
-);
