--- packages/web/src/components/Icons/Command.tsx
@@ -1,6 +1,0 @@
-import { Command, type IconProps } from "@phosphor-icons/react";
-import { getInteractiveIconClassName } from "./icon.utils";
-
-export const CommandIcon = ({ className, ...props }: IconProps) => (
-  <Command className={getInteractiveIconClassName(className)} {...props} />
-);
