--- packages/web/src/components/Icons/Flask.tsx
@@ -1,6 +1,0 @@
-import { Flask, type IconProps } from "@phosphor-icons/react";
-import { getInteractiveIconClassName } from "./icon.utils";
-
-export const FlaskIcon = ({ className, ...props }: IconProps) => (
-  <Flask className={getInteractiveIconClassName(className)} {...props} />
-);
