--- packages/web/src/components/Icons/List.tsx
@@ -1,11 +1,0 @@
-import { type IconProps, List } from "@phosphor-icons/react";
-import { getInteractiveIconClassName } from "./icon.utils";
-
-export const ListIcon = ({ className, ...props }: IconProps) => (
-  <List
-    className={getInteractiveIconClassName(
-      `cursor-pointer text-text-muted ${className ?? ""}`,
-    )}
-    {...props}
-  />
-);
