--- packages/web/src/components/Icons/Refresh.tsx
@@ -1,9 +1,0 @@
-import { ArrowsClockwiseIcon, type IconProps } from "@phosphor-icons/react";
-import { getInteractiveIconClassName } from "./icon.utils";
-
-export const RefreshIcon = ({ className, ...props }: IconProps) => (
-  <ArrowsClockwiseIcon
-    className={getInteractiveIconClassName(className)}
-    {...props}
-  />
-);
