+++ packages/web/src/components/Icons/Repeat.tsx
@@ -1,6 +1,12 @@
-import { type IconProps, Repeat } from "@phosphor-icons/react";
+import {
+  type IconProps,
+  RepeatIcon as PhosphorRepeatIcon,
+} from "@phosphor-icons/react";
 import { getInteractiveIconClassName } from "./icon.utils";
 
 export const RepeatIcon = ({ className, ...props }: IconProps) => (
-  <Repeat className={getInteractiveIconClassName(className)} {...props} />
+  <PhosphorRepeatIcon
+    className={getInteractiveIconClassName(className)}
+    {...props}
+  />
 );
