--- packages/web/src/components/Icons/Todo.tsx
@@ -1,6 +1,0 @@
-import { CheckCircle, type IconProps } from "@phosphor-icons/react";
-import { getInteractiveIconClassName } from "./icon.utils";
-
-export const TodoIcon = ({ className, ...props }: IconProps) => (
-  <CheckCircle className={getInteractiveIconClassName(className)} {...props} />
-);
