--- packages/web/src/components/Icons/icon.types.ts
@@ -1,5 +1,0 @@
-import { type IconProps as PhosphorIconProps } from "@phosphor-icons/react";
-
-export interface IconProps extends PhosphorIconProps {
-  ref?: React.Ref<SVGSVGElement>;
-}
