--- packages/web/src/components/Textarea/types.ts
@@ -1,12 +1,0 @@
-import { type TextareaAutosizeProps } from "react-textarea-autosize";
-import {
-  type ClassNamedComponent,
-  type UnderlinedInput,
-} from "@web/common/types/component.types";
-
-export interface TextareaProps
-  extends UnderlinedInput,
-    ClassNamedComponent,
-    TextareaAutosizeProps {
-  heightFitsContent?: boolean;
-}
